Replace json validator with newest version
