+++ composer.json
@@ -27,7 +27,7 @@
     "prooph/psr7-middleware": "^0.6",
     "psr/container": "^1.0",
     "zendframework/zend-diactoros": "^1.3",
-    "webmozart/json": "^1.2.2"
+    "justinrainbow/json-schema": "^5.2"
   },
   "require-dev": {
     "phpunit/phpunit": "^6.0"
